BAU-400 Move ChartBuilderState constituent types into chartBuilderReducer.ts
